:truck: Rename to project selector
