+++ src/dashboard/listItems.js
@@ -8,7 +8,7 @@ import DashboardIcon from "@material-ui/icons/Dashboard";
 import DnsIcon from "@material-ui/icons/Dns";
 import InfoIcon from "@material-ui/icons/Info";
 
-import ProjectSelect from "../components/selectProject";
+import ProjectSelect from "../components/ProjectSelector";
 
 export const projectList = (
   <div>
